[smg][ci] rename 3rd models from cloud backend and delete dead code
